Adding support for static assets and user info
Now requires that an assets flag be used on the command line to specify where
the web application will be hosted from.

Also now provides a UserInfo endpoint which will respond with a JSON formatted
object that contains the user profile from an OpenID Connect server, assuming
that the nginx gateway is set up in front of the application and is populating
the "X-USER" header.